requirements.txt
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
eventlet==0.40.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.19.0
mysql-connector==2.2.9
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
PyJWT==2.10.1
PyMySQL==1.0.2
PySocks==1.7.1
pytest==7.4.0
python-dotenv==1.0.0
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.4
selenium==4.33.0
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
